requirements.txt
# requirements.txt
Django==6.0
groq==0.37.1
python-dotenv==1.2.1
gunicorn==21.2.0
whitenoise==6.6.0
psycopg2-binary==2.9.9